refactor: ports contracts and admin logs split
