+++ backend/internal/application/audit/service.go
@@ -30,7 +30,7 @@ type Service struct {
 
 const (
 	defaultPageSize = 20
-	maxPageSize     = 100
+	maxPageSize     = 1000
 )
 
 // NewService 创建服务。
