+++ backend/internal/application/systemevent/service.go
@@ -42,7 +42,7 @@ type Service struct {
 
 const (
 	defaultPageSize = 20
-	maxPageSize     = 100
+	maxPageSize     = 1000
 )
 
 // NewService 创建系统事件服务。
