+++ backend/internal/ports/doc.go
@@ -1,0 +1,2 @@
+// Package ports 定义出站集成端口的数据契约：application 依赖这些类型声明消费方接口，infra 适配器以相同签名实现。
+package ports
