+++ backend/internal/transport/http/announcement/handler.go
@@ -260,7 +260,7 @@ func writeAnnouncementError(c *gin.Context, err error) {
 func pageParams(c *gin.Context) (int, int) {
 	page := 1
 	pageSize := 20
-	const maxPageSize = 100
+	const maxPageSize = 1000
 	if raw := c.Query("page"); raw != "" {
 		if parsed, err := strconv.Atoi(raw); err == nil && parsed > 0 {
 			page = parsed
