+++ backend/internal/transport/http/knowledgebase/handler.go
@@ -762,7 +762,7 @@ func listInput(c *gin.Context) appknowledgebase.ListInput {
 }
 
 func pageParams(c *gin.Context) (int, int) {
-	const maxPageSize = 100
+	const maxPageSize = 1000
 	page, _ := strconv.Atoi(c.DefaultQuery("page", "1"))
 	pageSize, _ := strconv.Atoi(c.DefaultQuery("page_size", "20"))
 	if page <= 0 {
