+++ backend/internal/transport/http/layering_test.go
@@ -65,6 +65,18 @@ func TestBackendLayeringImports(t *testing.T) {
 				`"github.com/DEEIX-AI/DEEIX-Chat/backend/internal/transport`,
 			},
 		},
+		{
+			dir: "ports",
+			forbidden: []string{
+				`"github.com/DEEIX-AI/DEEIX-Chat/backend/internal/application`,
+				`"github.com/DEEIX-AI/DEEIX-Chat/backend/internal/transport`,
+				`"github.com/DEEIX-AI/DEEIX-Chat/backend/internal/infra`,
+				`"github.com/DEEIX-AI/DEEIX-Chat/backend/internal/repository`,
+				`"github.com/gin-gonic/gin"`,
+				`"gorm.io/gorm"`,
+				`"github.com/redis/go-redis`,
+			},
+		},
 	}
 
 	for _, check := range checks {
@@ -72,6 +84,102 @@ func TestBackendLayeringImports(t *testing.T) {
 		t.Run(check.dir, func(t *testing.T) {
 			assertNoForbiddenImports(t, filepath.Join(root, check.dir), check.forbidden)
 		})
+	}
+}
+
+// TestApplicationInfraAdapterRatchet 冻结 application 对 infra 业务适配器的存量依赖：
+// 新增依赖直接失败；某个包还清依赖后必须同步删除对应白名单条目，确保依赖只减不增。
+// 还债方式见分层规范「出站集成端口」：数据契约放 internal/ports，接口由消费方定义，infra 按端口签名实现。
+func TestApplicationInfraAdapterRatchet(t *testing.T) {
+	const modulePrefix = "github.com/DEEIX-AI/DEEIX-Chat/backend/internal/"
+	// 规范允许的技术型组件：无业务决策、无持久化语义。
+	allowedTechnicalPrefixes := []string{
+		"infra/config",
+		"infra/observability",
+	}
+	// 存量业务适配器依赖白名单（application 包名 -> infra 包路径）。
+	allowlist := map[string][]string{
+		"auth":          {"infra/geoip", "infra/identityprovider"},
+		"channel":       {"infra/llm", "infra/objectstore"},
+		"compact":       {"infra/llm"},
+		"conversation":  {"infra/embedding", "infra/llm", "infra/mcp", "infra/objectstore"},
+		"embedding":     {"infra/embedding"},
+		"extraction":    {"infra/extract/builtin", "infra/extract/docling", "infra/extract/mineru", "infra/extract/ocr", "infra/extract/tika", "infra/objectstore"},
+		"mcp":           {"infra/mcp"},
+		"objectstorage": {"infra/objectstore"},
+		"rag":           {"infra/embedding"},
+		"runtime":       {"infra/extract/docling", "infra/extract/mineru", "infra/extract/ocr", "infra/extract/tika"},
+		"settings":      {"infra/extract/mineru"},
+		"upload":        {"infra/objectstore"},
+	}
+
+	applicationRoot := filepath.Clean("../../application")
+	actual := map[string]map[string]bool{}
+	fileSet := token.NewFileSet()
+	err := filepath.WalkDir(applicationRoot, func(path string, entry os.DirEntry, walkErr error) error {
+		if walkErr != nil || entry.IsDir() || filepath.Ext(path) != ".go" || strings.HasSuffix(path, "_test.go") {
+			return walkErr
+		}
+		relative, relErr := filepath.Rel(applicationRoot, path)
+		if relErr != nil {
+			return relErr
+		}
+		packageName := strings.Split(filepath.ToSlash(relative), "/")[0]
+		file, parseErr := parser.ParseFile(fileSet, path, nil, parser.ImportsOnly)
+		if parseErr != nil {
+			return parseErr
+		}
+		for _, importSpec := range file.Imports {
+			importPath, unquoteErr := strconv.Unquote(importSpec.Path.Value)
+			if unquoteErr != nil {
+				return unquoteErr
+			}
+			if !strings.HasPrefix(importPath, modulePrefix+"infra/") {
+				continue
+			}
+			infraPath := strings.TrimPrefix(importPath, modulePrefix)
+			if actual[packageName] == nil {
+				actual[packageName] = map[string]bool{}
+			}
+			actual[packageName][infraPath] = true
+		}
+		return nil
+	})
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	isAllowedTechnical := func(infraPath string) bool {
+		for _, prefix := range allowedTechnicalPrefixes {
+			if infraPath == prefix || strings.HasPrefix(infraPath, prefix+"/") {
+				return true
+			}
+		}
+		return false
+	}
+	inAllowlist := func(packageName, infraPath string) bool {
+		for _, item := range allowlist[packageName] {
+			if item == infraPath {
+				return true
+			}
+		}
+		return false
+	}
+
+	for packageName, imports := range actual {
+		for infraPath := range imports {
+			if isAllowedTechnical(infraPath) || inAllowlist(packageName, infraPath) {
+				continue
+			}
+			t.Errorf("application/%s 新增了 infra 业务适配器依赖 %q：请改为通过 internal/ports 定义端口契约，而不是扩大白名单", packageName, infraPath)
+		}
+	}
+	for packageName, entries := range allowlist {
+		for _, infraPath := range entries {
+			if !actual[packageName][infraPath] {
+				t.Errorf("application/%s 已不再依赖 %q：请从白名单删除该条目，锁定还债进度", packageName, infraPath)
+			}
+		}
 	}
 }
