+++ backend/internal/transport/http/promptpreset/handler.go
@@ -319,7 +319,7 @@ func boolQuery(c *gin.Context, key string) *bool {
 func pageParams(c *gin.Context) (int, int) {
 	page := 1
 	pageSize := 20
-	const maxPageSize = 100
+	const maxPageSize = 1000
 	if raw := c.Query("page"); raw != "" {
 		if parsed, err := strconv.Atoi(raw); err == nil && parsed > 0 {
 			page = parsed
