+++ frontend/features/admin/components/sections/logs/admin-logs.tsx
@@ -1,8 +1,8 @@
 "use client";
 
-import * as React from "react";
-import { CornerDownRight, Trash2 } from "lucide-react";
+import { Trash2 } from "lucide-react";
 import { useLocale, useTranslations } from "next-intl";
+import * as React from "react";
 
 import {
   AlertDialog,
@@ -24,15 +24,7 @@ import {
   SelectTrigger,
   SelectValue,
 } from "@/components/ui/select";
-import { Separator } from "@/components/ui/separator";
-import { Sheet, SheetContent, SheetDescription, SheetHeader, SheetTitle } from "@/components/ui/sheet";
-import { Spinner, SpinnerLabel } from "@/components/ui/spinner";
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipTrigger,
-} from "@/components/ui/tooltip";
+import { SpinnerLabel } from "@/components/ui/spinner";
 import {
   Table,
   TableBody,
@@ -43,12 +35,9 @@ import {
   TableLoadingRow,
   TableRow,
 } from "@/components/ui/table";
-import { useVirtualTableRows, VirtualTablePaddingRow } from "@/components/ui/virtual-table";
-import { AdminDateRangeFilter, ADMIN_DATE_PICKER_TRIGGER_CLASSNAME } from "@/features/admin/components/admin-date-range-filter";
-import { AdminDateTimePicker } from "@/features/admin/components/admin-date-time-picker";
 import { TablePagination, TableToolbar } from "@/components/ui/table-tools";
-import { CopyActionButton } from "@/shared/components/copy-action";
-import { useDialogSnapshot } from "@/shared/hooks/use-dialog-snapshot";
+import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
+import { useVirtualTableRows, VirtualTablePaddingRow } from "@/components/ui/virtual-table";
 import type {
   AdminAuditLogDTO,
   AdminConversationEventDTO,
@@ -57,1017 +46,52 @@ import type {
   AdminUserAuthEventDTO,
 } from "@/features/admin/api/admin.types";
 import { type AdminLogCleanupType } from "@/features/admin/api/audit";
+import { AdminDateRangeFilter } from "@/features/admin/components/admin-date-range-filter";
+import { AdminDateTimePicker } from "@/features/admin/components/admin-date-time-picker";
+import { LogDetailSheet } from "@/features/admin/components/sections/logs/admin-log-detail-sheet";
+import { ModerationEventTable } from "@/features/admin/components/sections/logs/admin-moderation-events";
+import {
+  UsageLogCostCell,
+  UsageLogModelCell,
+  UsageLogModelFilter,
+  UsageLogUsageCell,
+  useUsageBillingLabels,
+} from "@/features/admin/components/sections/logs/admin-usage-log-cells";
+import {
+  AUDIT_LOG_SORT_OPTIONS,
+  type AuditLogSortValue,
+  CONVERSATION_EVENT_SORT_OPTIONS,
+  type ConversationEventSortValue,
+  PAYMENT_ORDER_SORT_OPTIONS,
+  type PaymentOrderSortValue,
+  SECURITY_LOG_SORT_OPTIONS,
+  type SecurityLogSortValue,
+  USAGE_LOG_SORT_OPTIONS,
+  type UsageLogSortValue,
+  useAdminConversationEvents,
+  useAdminLogs,
+  useAdminPaymentOrders,
+  useAdminSecurityLogs,
+  useAdminUsageLogs,
+} from "@/features/admin/hooks/use-admin-logs";
 import {
   cleanupDateToISOString,
   useAdminBillingDisplayOptions,
   useAdminConversationRunsCleanup,
   useAdminLogCleanupDialog,
   useAdminLogDetail,
-  type LogDetail,
 } from "@/features/admin/hooks/use-admin-logs-actions";
 import {
-  AUDIT_LOG_SORT_OPTIONS,
-  CONVERSATION_EVENT_SORT_OPTIONS,
-  PAYMENT_ORDER_SORT_OPTIONS,
-  SECURITY_LOG_SORT_OPTIONS,
-  USAGE_LOG_SORT_OPTIONS,
-  useAdminConversationEvents,
-  useAdminLogs,
-  useAdminPaymentOrders,
-  useAdminSecurityLogs,
-  useAdminUsageLogs,
-  type AuditLogSortValue,
-  type ConversationEventSortValue,
-  type PaymentOrderSortValue,
-  type SecurityLogSortValue,
-  type UsageLogSortValue,
-} from "@/features/admin/hooks/use-admin-logs";
+  formatCount,
+  formatDateTime,
+  formatMoneyCents,
+  resolveUserDisplayName,
+} from "@/features/admin/model/log-display";
+import { formatUsageBalance } from "@/features/admin/model/usage-log-billing";
 import { cn } from "@/lib/utils";
-import { formatBillingBalance } from "@/features/admin/utils/account-display";
-import {
-  billingRateMultiplierNote,
-  cacheWriteBillingLabel,
-  cacheWriteBillingNote,
-  formatBillingDisplayCompactAmountFromUSD,
-  formatBillingDisplayPreciseAmountFromUSD,
-  formatBillingDisplayUnitPriceFromUSD,
-  type BillingDisplayLabels,
-  type BillingDisplayOptions,
-} from "@/shared/lib/billing-display";
-import { ModelSelect, type ModelSelectOption } from "@/shared/components/model-select";
-import { formatBytes } from "@/shared/lib/file-display";
-import { ModerationEventTable } from "@/features/admin/components/sections/logs/admin-moderation-events";
 import { useAuthSession } from "@/shared/auth/auth-session-context";
+import type { BillingDisplayOptions } from "@/shared/lib/billing-display";
 
-const ALL_MODELS_VALUE = "__all__";
-
-function formatDateTime(value: string | null | undefined, locale: string): string {
-  if (!value) {
-    return "-";
-  }
-  const date = new Date(value);
-  if (Number.isNaN(date.getTime())) {
-    return "-";
-  }
-  return new Intl.DateTimeFormat(locale, {
-    year: "numeric",
-    month: "2-digit",
-    day: "2-digit",
-    hour: "2-digit",
-    minute: "2-digit",
-  }).format(date);
-}
-
-function resolveUserDisplayName(label: string, username: string, fallbackID: number): string {
-  const name = label.trim() || username.trim();
-  return name || String(fallbackID);
-}
-
-function formatJSON(raw: string | null | undefined): string {
-  const value = raw?.trim();
-  if (!value) {
-    return "{}";
-  }
-  try {
-    return JSON.stringify(JSON.parse(value), null, 2);
-  } catch {
-    return value;
-  }
-}
-
-function parseJSONRecord(raw: string | null | undefined): Record<string, unknown> | null {
-  const value = raw?.trim();
-  if (!value) {
-    return null;
-  }
-  try {
-    const parsed = JSON.parse(value) as unknown;
-    return parsed && typeof parsed === "object" && !Array.isArray(parsed)
-      ? (parsed as Record<string, unknown>)
-      : null;
-  } catch {
-    return null;
-  }
-}
-
-function formatCount(value: number | null | undefined, locale: string): string {
-  return new Intl.NumberFormat(locale).format(value ?? 0);
-}
-
-function formatUsageBalance(value: number | null | undefined, billingDisplay: BillingDisplayOptions): string {
-  return value === null || value === undefined ? "-" : formatBillingBalance(value, billingDisplay);
-}
-
-function usageBillableOutputTokens(item: AdminUsageLogDTO): number {
-  return item.outputTokens + item.reasoningTokens;
-}
-
-function usageTotalTokens(item: AdminUsageLogDTO): number {
-  return item.inputTokens + item.cacheReadTokens + item.cacheWriteTokens + usageBillableOutputTokens(item);
-}
-
-type UsagePricingSnapshot = {
-  pricing_mode?: "token" | "call" | "duration" | "tiered" | string;
-  provider_protocol?: string;
-  duration_billable?: boolean;
-  media_type?: string;
-  input_image_count?: number;
-  cache_timeout?: string;
-  fast_mode?: boolean;
-  billing_speed?: string;
-  billing_service_tier?: string;
-  rate_multiplier?: number;
-  cache_write_5m_tokens?: number;
-  cache_write_1h_tokens?: number;
-  input_nanousd_per_m_tokens?: number;
-  cache_read_nanousd_per_m_tokens?: number;
-  cache_write_nanousd_per_m_tokens?: number;
-  output_nanousd_per_m_tokens?: number;
-  call_nanousd_per_call?: number;
-  duration_nanousd_per_second?: number;
-  input_billed_nanousd?: number;
-  cache_read_billed_nanousd?: number;
-  cache_write_billed_nanousd?: number;
-  output_billed_nanousd?: number;
-  call_billed_nanousd?: number;
-  duration_billed_nanousd?: number;
-  tiered_from_tokens?: number;
-  tiered_up_to_tokens?: number | null;
-  upstream_usage?: unknown;
-};
-
-function usageLogRawUsageJSON(item: AdminUsageLogDTO): string {
-  const upstreamUsage = parseJSONRecord(item.pricingSnapshotJSON)?.upstream_usage;
-  if (upstreamUsage && typeof upstreamUsage === "object") {
-    return JSON.stringify(upstreamUsage, null, 2);
-  }
-  return "{}";
-}
-
-type UsageBillingLabels = {
-  input: string;
-  output: string;
-  cacheRead: string;
-  total: string;
-  freeModelNoBilling: string;
-  perCall: string;
-  perSecond: string;
-  callUnit: string;
-  secondUnit: string;
-  rateNote: string;
-  cacheNote: string;
-  tieredRangeBounded: (from: string, upTo: string) => string;
-  tieredRangeOpen: (from: string) => string;
-  table: {
-    item: string;
-    usage: string;
-    unitPrice: string;
-    amount: string;
-  };
-  billingDisplay: BillingDisplayLabels;
-};
-
-function useUsageBillingLabels(): UsageBillingLabels {
-  const t = useTranslations("adminLogs.usage.billing");
-
-  return React.useMemo(
-    () => ({
-      input: t("input"),
-      output: t("output"),
-      cacheRead: t("cacheRead"),
-      total: t("total"),
-      freeModelNoBilling: t("freeModelNoBilling"),
-      perCall: t("perCall"),
-      perSecond: t("perSecond"),
-      callUnit: t("callUnit"),
-      secondUnit: t("secondUnit"),
-      rateNote: t("rateNote"),
-      cacheNote: t("cacheNote"),
-      tieredRangeBounded: (from: string, upTo: string) => t("tieredRangeBounded", { from, upTo }),
-      tieredRangeOpen: (from: string) => t("tieredRangeOpen", { from }),
-      table: {
-        item: t("table.item"),
-        usage: t("table.usage"),
-        unitPrice: t("table.unitPrice"),
-        amount: t("table.amount"),
-      },
-      billingDisplay: {
-        cacheWrite: t("cacheWrite"),
-        cacheWrite5m: t("cacheWrite5m"),
-        cacheWrite1h: t("cacheWrite1h"),
-        cacheWrite5m1h: t("cacheWrite5m1h"),
-        cacheWritePricingLabel: t("cacheWritePricingLabel"),
-        cacheWritePricingNote: t("cacheWritePricingNote"),
-        claudeCacheWriteMixedNote: (multiplier: string) => t("claudeCacheWriteMixedNote", { multiplier }),
-        claudeCacheWriteNote: (timeout: "5m" | "1h", multiplier: string) => t("claudeCacheWriteNote", { timeout, multiplier }),
-        claudeFastModeNote: (multiplier: string) => t("claudeFastModeNote", { multiplier }),
-        openaiServiceTierNote: (tier: string, multiplier: string) => t("openaiServiceTierNote", { tier, multiplier }),
-      },
-    }),
-    [t],
-  );
-}
-
-function formatUsageCost(value: number, billingDisplay: BillingDisplayOptions): string {
-  return formatBillingDisplayCompactAmountFromUSD(value, billingDisplay);
-}
-
-function formatTooltipUsageCost(value: number, billingDisplay: BillingDisplayOptions): string {
-  return formatBillingDisplayPreciseAmountFromUSD(value, billingDisplay);
-}
-
-function formatMoneyCents(value: number | null | undefined, currency: string): string {
-  const amount = (value ?? 0) / 100;
-  const normalizedCurrency = currency.trim().toUpperCase();
-  if (!normalizedCurrency) {
-    return amount.toLocaleString("en-US", { minimumFractionDigits: 2, maximumFractionDigits: 2 });
-  }
-  try {
-    return new Intl.NumberFormat("en-US", {
-      style: "currency",
-      currency: normalizedCurrency,
-      minimumFractionDigits: 2,
-      maximumFractionDigits: 2,
-    }).format(amount);
-  } catch {
-    return `${amount.toLocaleString("en-US", { minimumFractionDigits: 2, maximumFractionDigits: 2 })} ${normalizedCurrency}`;
-  }
-}
-
-function formatTooltipUnitPrice(value: number, billingDisplay: BillingDisplayOptions): string {
-  return formatBillingDisplayUnitPriceFromUSD(value, billingDisplay);
-}
-
-function nanousdToUSD(value: number): number {
-  if (!Number.isFinite(value) || value <= 0) return 0;
-  return value / 1_000_000_000;
-}
-
-function parseUsagePricingSnapshot(raw: string): UsagePricingSnapshot {
-  try {
-    const parsed = JSON.parse(raw) as unknown;
-    return parsed && typeof parsed === "object" && !Array.isArray(parsed) ? parsed as UsagePricingSnapshot : {};
-  } catch {
-    return {};
-  }
-}
-
-function readUsageSnapshotNumber(snapshot: UsagePricingSnapshot, key: keyof UsagePricingSnapshot): number {
-  const value = snapshot[key];
-  return typeof value === "number" && Number.isFinite(value) ? value : 0;
-}
-
-function normalizePricingMode(value: string | null | undefined): "token" | "call" | "duration" | "tiered" {
-  if (value === "call" || value === "duration" || value === "tiered") return value;
-  return "token";
-}
-
-function calcTokenBilledNanousd(tokens: number, rateNanousd: number): number {
-  if (!Number.isFinite(tokens) || !Number.isFinite(rateNanousd) || tokens <= 0 || rateNanousd <= 0) return 0;
-  return Math.round((tokens * rateNanousd) / 1_000_000);
-}
-
-function resolveTokenBilledNanousd(snapshot: UsagePricingSnapshot, billedKey: keyof UsagePricingSnapshot, tokens: number, rateNanousd: number): number {
-  const billed = readUsageSnapshotNumber(snapshot, billedKey);
-  return billed > 0 ? billed : calcTokenBilledNanousd(tokens, rateNanousd);
-}
-
-function resolveCountBilledNanousd(snapshot: UsagePricingSnapshot, billedKey: keyof UsagePricingSnapshot, count: number, rateNanousd: number): number {
-  const billed = readUsageSnapshotNumber(snapshot, billedKey);
-  if (billed > 0) return billed;
-  if (!Number.isFinite(count) || !Number.isFinite(rateNanousd) || count <= 0 || rateNanousd <= 0) return 0;
-  return Math.round(count * rateNanousd);
-}
-
-function formatFormulaTokenCount(value: number): string {
-  if (!Number.isFinite(value) || value <= 0) return "0";
-  return value.toLocaleString("en-US");
-}
-
-function formatTieredRangeLabel(fromTokens: number | null | undefined, upToTokens: number | null | undefined, labels: UsageBillingLabels): string {
-  const from = Number.isFinite(fromTokens ?? NaN) && (fromTokens ?? 0) > 0 ? fromTokens ?? 0 : 0;
-  const upTo = Number.isFinite(upToTokens ?? NaN) && (upToTokens ?? 0) > 0 ? upToTokens ?? 0 : null;
-  return upTo
-    ? labels.tieredRangeBounded(formatFormulaTokenCount(from), formatFormulaTokenCount(upTo))
-    : labels.tieredRangeOpen(formatFormulaTokenCount(from));
-}
-
-type UsageBillingTooltipLine =
-  | { type: "row"; left: string; right: string }
-  | { type: "divider" }
-  | { type: "tiered-table"; rangeLabel: string; rows: UsageBillingTieredTableRow[]; totalLabel: string; totalAmount: string };
-
-type UsageBillingTieredTableRow = {
-  item: string;
-  tokens: string;
-  unitPrice: string;
-  amount: string;
-};
-
-function UsageBillingTooltipLines({ lines, labels }: { lines: UsageBillingTooltipLine[]; labels: UsageBillingLabels }) {
-  return (
-    <div className="min-w-72 max-w-[min(92vw,44rem)] space-y-1 text-left text-xs leading-relaxed">
-      {lines.map((line, index) =>
-        line.type === "divider" ? (
-          <Separator key={`divider-${index}`} />
-        ) : line.type === "tiered-table" ? (
-          <UsageBillingTieredTable key={`tiered-table-${index}`} line={line} labels={labels} />
-        ) : (
-          <div key={`${line.left}-${index}`} className="grid grid-cols-[minmax(0,1fr)_auto] items-baseline gap-8">
-            <span className="min-w-0 text-left">{line.left}</span>
-            <span className="whitespace-nowrap text-right tabular-nums">{line.right}</span>
-          </div>
-        ),
-      )}
-    </div>
-  );
-}
-
-function UsageBillingTieredTable({
-  line,
-  labels,
-}: {
-  line: Extract<UsageBillingTooltipLine, { type: "tiered-table" }>;
-  labels: UsageBillingLabels;
-}) {
-  return (
-    <div className="max-w-[min(92vw,34rem)] overflow-x-auto">
-      <div className="mb-1 text-[11px] font-medium text-background/80">{line.rangeLabel}</div>
-      <table className="w-full border-collapse text-left tabular-nums">
-        <thead>
-          <tr className="border-b border-background/20 text-[11px] text-background/65">
-            <th className="whitespace-nowrap px-2 pb-1 font-medium first:pl-0" aria-label={labels.table.item} />
-            <th className="whitespace-nowrap px-2 pb-1 text-right font-medium">{labels.table.usage}</th>
-            <th className="whitespace-nowrap px-2 pb-1 text-right font-medium">{labels.table.unitPrice}</th>
-            <th className="whitespace-nowrap px-2 pb-1 text-right font-medium last:pr-0">{labels.table.amount}</th>
-          </tr>
-        </thead>
-        <tbody>
-          {line.rows.map((row, rowIndex) => (
-            <tr key={`${row.item}-${rowIndex}`} className="border-b border-background/10 last:border-0">
-              <td className="whitespace-nowrap px-2 py-1 first:pl-0">{row.item}</td>
-              <td className="whitespace-nowrap px-2 py-1 text-right">{row.tokens}</td>
-              <td className="whitespace-nowrap px-2 py-1 text-right">{row.unitPrice}</td>
-              <td className="whitespace-nowrap px-2 py-1 text-right last:pr-0">{row.amount}</td>
-            </tr>
-          ))}
-        </tbody>
-        <tfoot>
-          <tr className="border-t border-background/20">
-            <td className="px-2 pt-1.5 font-medium first:pl-0" colSpan={3}>{line.totalLabel}</td>
-            <td className="whitespace-nowrap px-2 pt-1.5 text-right font-medium last:pr-0">{line.totalAmount}</td>
-          </tr>
-        </tfoot>
-      </table>
-    </div>
-  );
-}
-
-function usageFormulaLine(
-  label: string,
-  tokens: number,
-  rateNanousd: number,
-  billedNanousd: number,
-  billingDisplay: BillingDisplayOptions,
-): UsageBillingTooltipLine {
-  return {
-    type: "row",
-    left: label,
-    right: `${formatFormulaTokenCount(tokens)} tokens * ${formatTooltipUnitPrice(nanousdToUSD(rateNanousd), billingDisplay)} / 1M = ${formatTooltipUsageCost(nanousdToUSD(billedNanousd), billingDisplay)}`,
-  };
-}
-
-function usageCountFormulaLine(
-  label: string,
-  count: number,
-  unit: string,
-  rateUnit: string,
-  rateNanousd: number,
-  billedNanousd: number,
-  billingDisplay: BillingDisplayOptions,
-): UsageBillingTooltipLine {
-  const safeCount = Number.isFinite(count) && count > 0 ? count : 0;
-  return {
-    type: "row",
-    left: label,
-    right: `${safeCount.toLocaleString("en-US")} ${unit} * ${formatTooltipUnitPrice(nanousdToUSD(rateNanousd), billingDisplay)} / ${rateUnit} = ${formatTooltipUsageCost(nanousdToUSD(billedNanousd), billingDisplay)}`,
-  };
-}
-
-function usageTieredTableRow(
-  item: string,
-  tokens: number,
-  rateNanousd: number,
-  billedNanousd: number,
-  billingDisplay: BillingDisplayOptions,
-): UsageBillingTieredTableRow {
-  const safeTokens = Number.isFinite(tokens) && tokens > 0 ? tokens : 0;
-  const safeBilled = Number.isFinite(billedNanousd) && billedNanousd > 0 ? billedNanousd : 0;
-  return {
-    item,
-    tokens: formatFormulaTokenCount(safeTokens),
-    unitPrice: `${formatTooltipUnitPrice(nanousdToUSD(rateNanousd), billingDisplay)} / 1M`,
-    amount: formatTooltipUsageCost(nanousdToUSD(safeBilled), billingDisplay),
-  };
-}
-
-function usageTotalLine(item: AdminUsageLogDTO, labels: UsageBillingLabels, billingDisplay: BillingDisplayOptions): UsageBillingTooltipLine {
-  return {
-    type: "row",
-    left: labels.total,
-    right: item.isFreeModel
-      ? `${formatTooltipUsageCost(0, billingDisplay)} (${labels.freeModelNoBilling})`
-      : formatTooltipUsageCost(nanousdToUSD(item.billedNanousd), billingDisplay),
-  };
-}
-
-function buildUsageBillingTooltipLines(
-  item: AdminUsageLogDTO,
-  labels: UsageBillingLabels,
-  billingDisplay: BillingDisplayOptions,
-): UsageBillingTooltipLine[] {
-  const snapshot = parseUsagePricingSnapshot(item.pricingSnapshotJSON);
-  const pricingMode = normalizePricingMode(snapshot.pricing_mode);
-  const inputRate = readUsageSnapshotNumber(snapshot, "input_nanousd_per_m_tokens");
-  const outputRate = readUsageSnapshotNumber(snapshot, "output_nanousd_per_m_tokens");
-  const cacheReadRate = readUsageSnapshotNumber(snapshot, "cache_read_nanousd_per_m_tokens");
-  const cacheWriteRate = readUsageSnapshotNumber(snapshot, "cache_write_nanousd_per_m_tokens");
-  const billedOutputTokens = usageBillableOutputTokens(item);
-  const totalLine = usageTotalLine(item, labels, billingDisplay);
-  const cacheWriteLabel = cacheWriteBillingLabel(snapshot, labels.billingDisplay);
-  const cacheWriteNote = cacheWriteBillingNote(snapshot, labels.billingDisplay);
-  const rateMultiplierNote = billingRateMultiplierNote(snapshot, labels.billingDisplay);
-
-  if (pricingMode === "call") {
-    const callRate = readUsageSnapshotNumber(snapshot, "call_nanousd_per_call");
-    const callBilled = resolveCountBilledNanousd(snapshot, "call_billed_nanousd", item.callCount, callRate);
-    return [
-      usageCountFormulaLine(labels.perCall, item.callCount, labels.callUnit, labels.callUnit, callRate, callBilled, billingDisplay),
-      { type: "divider" },
-      totalLine,
-    ];
-  }
-
-  if (pricingMode === "duration") {
-    const durationRate = readUsageSnapshotNumber(snapshot, "duration_nanousd_per_second");
-    const durationBilled = resolveCountBilledNanousd(snapshot, "duration_billed_nanousd", item.durationSeconds, durationRate);
-    return [
-      usageCountFormulaLine(labels.perSecond, item.durationSeconds, labels.secondUnit, labels.secondUnit, durationRate, durationBilled, billingDisplay),
-      { type: "divider" },
-      totalLine,
-    ];
-  }
-
-  if (pricingMode === "tiered") {
-    const lines: UsageBillingTooltipLine[] = [];
-    if (rateMultiplierNote || cacheWriteNote) {
-      if (rateMultiplierNote) {
-        lines.push({ type: "row", left: labels.rateNote, right: rateMultiplierNote });
-      }
-      if (cacheWriteNote) {
-        lines.push({ type: "row", left: labels.cacheNote, right: cacheWriteNote });
-      }
-      lines.push({ type: "divider" });
-    }
-    lines.push({
-      type: "tiered-table",
-      rangeLabel: formatTieredRangeLabel(snapshot.tiered_from_tokens, snapshot.tiered_up_to_tokens, labels),
-      rows: [
-        usageTieredTableRow(labels.input, item.inputTokens, inputRate, readUsageSnapshotNumber(snapshot, "input_billed_nanousd"), billingDisplay),
-        usageTieredTableRow(labels.output, billedOutputTokens, outputRate, readUsageSnapshotNumber(snapshot, "output_billed_nanousd"), billingDisplay),
-        usageTieredTableRow(labels.cacheRead, item.cacheReadTokens, cacheReadRate, readUsageSnapshotNumber(snapshot, "cache_read_billed_nanousd"), billingDisplay),
-        usageTieredTableRow(cacheWriteLabel, item.cacheWriteTokens, cacheWriteRate, readUsageSnapshotNumber(snapshot, "cache_write_billed_nanousd"), billingDisplay),
-      ],
-      totalLabel: labels.total,
-      totalAmount: item.isFreeModel
-        ? `${formatTooltipUsageCost(0, billingDisplay)} (${labels.freeModelNoBilling})`
-        : formatTooltipUsageCost(nanousdToUSD(item.billedNanousd), billingDisplay),
-    });
-    return lines;
-  }
-
-  const inputBilled = resolveTokenBilledNanousd(snapshot, "input_billed_nanousd", item.inputTokens, inputRate);
-  const cacheReadBilled = resolveTokenBilledNanousd(snapshot, "cache_read_billed_nanousd", item.cacheReadTokens, cacheReadRate);
-  const cacheWriteBilled = resolveTokenBilledNanousd(snapshot, "cache_write_billed_nanousd", item.cacheWriteTokens, cacheWriteRate);
-  const outputBilled = resolveTokenBilledNanousd(snapshot, "output_billed_nanousd", billedOutputTokens, outputRate);
-  const lines: UsageBillingTooltipLine[] = [
-    usageFormulaLine(labels.input, item.inputTokens, inputRate, inputBilled, billingDisplay),
-    usageFormulaLine(labels.output, billedOutputTokens, outputRate, outputBilled, billingDisplay),
-    usageFormulaLine(labels.cacheRead, item.cacheReadTokens, cacheReadRate, cacheReadBilled, billingDisplay),
-    usageFormulaLine(cacheWriteLabel, item.cacheWriteTokens, cacheWriteRate, cacheWriteBilled, billingDisplay),
-    { type: "divider" },
-    totalLine,
-  ];
-  const noteLines: UsageBillingTooltipLine[] = [];
-  if (rateMultiplierNote) {
-    noteLines.push({ type: "row", left: labels.rateNote, right: rateMultiplierNote });
-  }
-  if (cacheWriteNote) {
-    noteLines.push({ type: "row", left: labels.cacheNote, right: cacheWriteNote });
-  }
-  if (noteLines.length > 0) {
-    lines.splice(4, 0, ...noteLines);
-  }
-  return lines;
-}
-
-function UsageLogModelCell({ item, labels }: { item: AdminUsageLogDTO; labels: UsageBillingLabels }) {
-  const t = useTranslations("adminLogs.usage.modelTooltip");
-  const lines: UsageBillingTooltipLine[] = [
-    { type: "row", left: t("upstreamName"), right: item.upstreamName || "-" },
-    { type: "row", left: t("upstreamModel"), right: item.upstreamModelName || "-" },
-    { type: "row", left: t("bindingCode"), right: item.routedBindingCode || "-" },
-    { type: "row", left: t("protocol"), right: item.providerProtocol || "-" },
-  ];
-
-  return (
-    <Tooltip>
-      <TooltipTrigger asChild>
-        <div className="grid min-w-0 cursor-default gap-px">
-          <div className="max-w-[15rem] truncate font-medium leading-4" title={item.platformModelName || "-"}>
-            {item.platformModelName || "-"}
-          </div>
-          <div className="flex min-w-0 items-center gap-1 font-mono leading-4 text-muted-foreground">
-            <CornerDownRight className="size-3 shrink-0 stroke-1" />
-            <span className="max-w-[14rem] truncate" title={item.upstreamModelName || "-"}>
-              {item.upstreamModelName || "-"}
-            </span>
-          </div>
-        </div>
-      </TooltipTrigger>
-      <TooltipContent side="top">
-        <UsageBillingTooltipLines lines={lines} labels={labels} />
-      </TooltipContent>
-    </Tooltip>
-  );
-}
-
-function UsageLogUsageCell({ item, locale }: { item: AdminUsageLogDTO; locale: string }) {
-  const t = useTranslations("adminLogs.usage.tokens");
-  const snapshot = parseUsagePricingSnapshot(item.pricingSnapshotJSON);
-  const isVideoUsage = snapshot.media_type === "video" || snapshot.duration_billable === true;
-  if (isVideoUsage) {
-    const inputImageCount = typeof snapshot.input_image_count === "number" && Number.isFinite(snapshot.input_image_count) && snapshot.input_image_count >= 0
-      ? Math.trunc(snapshot.input_image_count)
-      : null;
-    const mediaUsage = [
-      { label: t("input"), value: inputImageCount === null ? "—" : t("imageCount", { count: inputImageCount }) },
-      { label: t("output"), value: t("secondCount", { count: item.durationSeconds }) },
-    ];
-    return (
-      <div className="grid min-w-[10.5rem] gap-1">
-        {mediaUsage.map((entry) => (
-          <span
-            key={entry.label}
-            className="inline-flex h-5 items-center justify-between gap-2 rounded-md bg-muted/45 px-1.5 text-[11px] leading-none text-muted-foreground"
-          >
-            <span>{entry.label}</span>
-            <span className="font-mono tabular-nums">{entry.value}</span>
-          </span>
-        ))}
-      </div>
-    );
-  }
-  const tokens = [
-    { label: t("inputShort"), value: item.inputTokens },
-    {
-      label: t("outputShort"),
-      value: usageBillableOutputTokens(item),
-      breakdown: {
-        visible: item.outputTokens,
-        reasoning: item.reasoningTokens,
-      },
-    },
-    { label: t("cacheReadShort"), value: item.cacheReadTokens },
-    { label: t("cacheWriteShort"), value: item.cacheWriteTokens },
-  ];
-
-  return (
-    <div className="grid min-w-[10.5rem] grid-cols-2 gap-1">
-      {tokens.map((token) => {
-        const badge = (
-          <span className={cn(
-            "inline-flex h-5 items-center justify-between gap-1 rounded-md bg-muted/45 px-1.5 font-mono text-[11px] leading-none text-muted-foreground",
-            token.breakdown && "cursor-help",
-          )}>
-            <span>{token.label}</span>
-            <span className="tabular-nums">{formatCount(token.value, locale)}</span>
-          </span>
-        );
-        if (!token.breakdown) {
-          return <React.Fragment key={token.label}>{badge}</React.Fragment>;
-        }
-        return (
-          <Tooltip key={token.label}>
-            <TooltipTrigger asChild>{badge}</TooltipTrigger>
-            <TooltipContent side="top" className="min-w-40">
-              <div className="grid gap-1.5 text-xs">
-                <div className="flex items-center justify-between gap-5">
-                  <span>{t("output")}</span>
-                  <span className="font-mono tabular-nums">{formatCount(token.breakdown.visible, locale)}</span>
-                </div>
-                <div className="flex items-center justify-between gap-5">
-                  <span>{t("reasoning")}</span>
-                  <span className="font-mono tabular-nums">{formatCount(token.breakdown.reasoning, locale)}</span>
-                </div>
-                <Separator className="bg-background/20" />
-                <div className="flex items-center justify-between gap-5 font-medium">
-                  <span>{t("outputTotal")}</span>
-                  <span className="font-mono tabular-nums">{formatCount(token.value, locale)}</span>
-                </div>
-              </div>
-            </TooltipContent>
-          </Tooltip>
-        );
-      })}
-    </div>
-  );
-}
-
-function UsageLogCostCell({
-  item,
-  labels,
-  billingDisplay,
-}: {
-  item: AdminUsageLogDTO;
-  labels: UsageBillingLabels;
-  billingDisplay: BillingDisplayOptions;
-}) {
-  const lines = buildUsageBillingTooltipLines(item, labels, billingDisplay);
-
-  return (
-    <Tooltip>
-      <TooltipTrigger asChild>
-        <span className={cn("inline-flex cursor-default items-center font-medium tabular-nums", item.isFreeModel ? "text-muted-foreground" : "text-foreground")}>
-          {item.isFreeModel ? labels.freeModelNoBilling : formatUsageCost(item.billedUSD, billingDisplay)}
-        </span>
-      </TooltipTrigger>
-      <TooltipContent side="top">
-        <UsageBillingTooltipLines lines={lines} labels={labels} />
-      </TooltipContent>
-    </Tooltip>
-  );
-}
-
-function UsageLogModelFilter({
-  value,
-  options,
-  disabled,
-  onChange,
-}: {
-  value: string;
-  options: ModelSelectOption[];
-  disabled: boolean;
-  onChange: (value: string) => void;
-}) {
-  const t = useTranslations("adminLogs.usage.filters");
-  const allOption = React.useMemo<ModelSelectOption>(() => ({ label: t("allModels"), value: ALL_MODELS_VALUE, iconUrl: null }), [t]);
-  const modelOptions = React.useMemo(() => [allOption, ...options], [allOption, options]);
-
-  return (
-    <ModelSelect
-      value={value.trim() || ALL_MODELS_VALUE}
-      fallbackValue={ALL_MODELS_VALUE}
-      disabled={disabled}
-      options={modelOptions}
-      align="start"
-      valueAlign="start"
-      itemAlign="start"
-      contentClassName="min-w-[320px]"
-      triggerClassName={cn(ADMIN_DATE_PICKER_TRIGGER_CLASSNAME, "h-7 px-2.5 text-[11px]")}
-      valueClassName={!value.trim() ? "text-muted-foreground" : undefined}
-      onChange={(nextValue) => onChange(nextValue === ALL_MODELS_VALUE ? "" : nextValue)}
-    />
-  );
-}
-
-function DetailRow({ label, value, mono = false }: { label: string; value: React.ReactNode; mono?: boolean }) {
-  return (
-    <div className="grid grid-cols-[88px_minmax(0,1fr)] gap-3 border-b border-border/50 py-2.5 last:border-b-0">
-      <p className="text-xs text-muted-foreground">{label}</p>
-      <div className={cn("min-w-0 break-words text-xs leading-5 text-foreground/86", mono && "font-mono")}>{value ?? "-"}</div>
-    </div>
-  );
-}
-
-function DetailBlock({ title, children }: { title: string; children: React.ReactNode }) {
-  return (
-    <section className="space-y-2">
-      <h4 className="px-1 text-xs font-medium text-foreground/88">{title}</h4>
-      <div className="rounded-lg border border-border/60 bg-background px-3">{children}</div>
-    </section>
-  );
-}
-
-function LogDetailSheet({
-  detail: rawDetail,
-  billingDisplay,
-  conversationDetailLoading,
-  onClose,
-}: {
-  detail: LogDetail | null;
-  billingDisplay: BillingDisplayOptions;
-  conversationDetailLoading: boolean;
-  onClose: () => void;
-}) {
-  const locale = useLocale();
-  const t = useTranslations("adminLogs.detail");
-  const usageLabels = useUsageBillingLabels();
-  const detail = useDialogSnapshot(rawDetail);
-  const copyMessages = React.useMemo(() => ({
-    copied: t("copied", { label: "" }).trim(),
-    failed: t("copyFailed"),
-  }), [t]);
-  const resultLabel = React.useCallback(
-    (value: string) => {
-      switch (value) {
-        case "success":
-          return t("result.success");
-        case "failure":
-          return t("result.failure");
-        case "blocked":
-          return t("result.blocked");
-        default:
-          return value || "-";
-      }
-    },
-    [t],
-  );
-  const title =
-    detail?.kind === "auth"
-      ? t("titles.auth")
-      : detail?.kind === "usage"
-        ? t("titles.usage")
-        : detail?.kind === "order"
-          ? t("titles.order")
-          : detail?.kind === "conversation"
-            ? t("titles.conversation")
-        : detail?.kind === "system"
-          ? t("titles.system")
-          : t("titles.audit");
-  const description =
-    detail?.kind === "auth"
-      ? `${detail.item.eventType || t("fallbacks.authEvent")} · ${formatDateTime(detail.item.occurredAt, locale)}`
-      : detail?.kind === "usage"
-        ? `${detail.item.platformModelName || t("fallbacks.modelCall")} · ${formatDateTime(detail.item.createdAt, locale)}`
-        : detail?.kind === "order"
-          ? `${detail.item.orderNo || t("fallbacks.order")} · ${formatDateTime(detail.item.createdAt, locale)}`
-          : detail?.kind === "conversation"
-            ? `${detail.item.eventType || detail.item.eventScope || t("fallbacks.conversationEvent")} · ${formatDateTime(detail.item.createdAt, locale)}`
-      : detail?.kind === "system"
-        ? `${detail.item.event || t("fallbacks.systemEvent")} · ${formatDateTime(detail.item.createdAt, locale)}`
-        : `${detail?.item.action || t("fallbacks.auditEvent")} · ${formatDateTime(detail?.item.createdAt, locale)}`;
-  const requestID = detail && detail.kind !== "usage" && detail.kind !== "order" && detail.kind !== "conversation" ? detail.item.requestID : "";
-  const detailJSON =
-    detail?.kind === "usage"
-      ? detail.item.pricingSnapshotJSON
-      : detail?.kind === "order"
-        ? detail.item.snapshotJSON
-        : detail?.kind === "conversation"
-          ? detail.item.payloadJSON || detail.item.inputJSON || detail.item.outputJSON || detail.item.errorJSON
-          : detail?.item.detailJSON;
-  const rawUsageJSON = detail?.kind === "usage" ? usageLogRawUsageJSON(detail.item) : "";
-  const formattedJSON = formatJSON(detailJSON);
-
-  return (
-    <Sheet open={Boolean(rawDetail)} onOpenChange={(open) => !open && onClose()}>
-      <SheetContent className="sm:max-w-[480px]">
-        <SheetHeader>
-          <SheetTitle>{title}</SheetTitle>
-          <SheetDescription>{description}</SheetDescription>
-        </SheetHeader>
-
-        <div className="min-h-0 flex-1 space-y-5 overflow-y-auto px-6 pb-6">
-          {detail?.kind === "audit" ? (
-            <>
-              <DetailBlock title={t("blocks.event")}>
-                <DetailRow label="ID" value={detail.item.id} mono />
-                <DetailRow label={t("fields.action")} value={detail.item.action} />
-                <DetailRow label={t("fields.resource")} value={detail.item.resource} />
-                <DetailRow label={t("fields.resourceID")} value={detail.item.resourceID} mono />
-                <DetailRow label={t("fields.createdAt")} value={formatDateTime(detail.item.createdAt, locale)} />
-              </DetailBlock>
-              <DetailBlock title={t("blocks.actor")}>
-                <DetailRow label={t("fields.user")} value={resolveUserDisplayName(detail.item.actorLabel, detail.item.actorUsername, detail.item.actorUserID)} />
-                <DetailRow label={t("fields.userID")} value={detail.item.actorUserID} mono />
-              </DetailBlock>
-              <DetailBlock title={t("blocks.request")}>
-                <DetailRow label={t("fields.requestID")} value={detail.item.requestID} mono />
-                <DetailRow label="IP" value={detail.item.ip} mono />
-                <DetailRow label="User Agent" value={detail.item.userAgent} />
-              </DetailBlock>
-            </>
-          ) : null}
-
-          {detail?.kind === "auth" ? (
-            <>
-              <DetailBlock title={t("blocks.event")}>
-                <DetailRow label="ID" value={detail.item.id} mono />
-                <DetailRow label={t("fields.event")} value={detail.item.eventType} />
-                <DetailRow label={t("fields.result")} value={resultLabel(detail.item.result)} />
-                <DetailRow label={t("fields.reason")} value={detail.item.reason} />
-                <DetailRow label={t("fields.occurredAt")} value={formatDateTime(detail.item.occurredAt, locale)} />
-              </DetailBlock>
-              <DetailBlock title={t("blocks.user")}>
-                <DetailRow label={t("fields.user")} value={resolveUserDisplayName(detail.item.userLabel, detail.item.username, detail.item.userID)} />
-                <DetailRow label={t("fields.userID")} value={detail.item.userID} mono />
-              </DetailBlock>
-              <DetailBlock title={t("blocks.request")}>
-                <DetailRow label={t("fields.requestID")} value={detail.item.requestID} mono />
-                <DetailRow label="IP" value={detail.item.clientIP} mono />
-                <DetailRow label="User Agent" value={detail.item.userAgent} />
-              </DetailBlock>
-            </>
-          ) : null}
-
-          {detail?.kind === "system" ? (
-            <>
-              <DetailBlock title={t("blocks.event")}>
-                <DetailRow label="ID" value={detail.item.id} mono />
-                <DetailRow label={t("fields.level")} value={detail.item.level} />
-                <DetailRow label={t("fields.source")} value={detail.item.source} />
-                <DetailRow label={t("fields.event")} value={detail.item.event} />
-                <DetailRow label={t("fields.message")} value={detail.item.message} />
-                <DetailRow label={t("fields.createdAt")} value={formatDateTime(detail.item.createdAt, locale)} />
-              </DetailBlock>
-              <DetailBlock title={t("blocks.resource")}>
-                <DetailRow label={t("fields.resource")} value={detail.item.resource} />
-                <DetailRow label={t("fields.resourceID")} value={detail.item.resourceID} mono />
-              </DetailBlock>
-              <DetailBlock title={t("blocks.request")}>
-                <DetailRow label={t("fields.requestID")} value={detail.item.requestID} mono />
-                <DetailRow label="Trace ID" value={detail.item.traceID} mono />
-              </DetailBlock>
-            </>
-          ) : null}
-
-          {detail?.kind === "usage" ? (
-            <>
-              <DetailBlock title={t("blocks.call")}>
-                <DetailRow label="ID" value={detail.item.id} mono />
-                <DetailRow label={t("fields.caller")} value={resolveUserDisplayName(detail.item.userLabel, detail.item.username, detail.item.userID)} />
-                <DetailRow label={t("fields.userID")} value={detail.item.userID} mono />
-                <DetailRow label={t("fields.conversationID")} value={detail.item.conversationID} mono />
-                <DetailRow label={t("fields.callTime")} value={formatDateTime(detail.item.createdAt, locale)} />
-              </DetailBlock>
-              <DetailBlock title={t("blocks.modelRoute")}>
-                <DetailRow label={t("fields.platformModel")} value={detail.item.platformModelName} mono />
-                <DetailRow label={t("fields.upstreamName")} value={detail.item.upstreamName} />
-                <DetailRow label={t("fields.upstreamModel")} value={detail.item.upstreamModelName} mono />
-                <DetailRow label={t("fields.bindingCode")} value={detail.item.routedBindingCode} mono />
-                <DetailRow label={t("fields.protocol")} value={detail.item.providerProtocol} />
-              </DetailBlock>
-              <DetailBlock title={t("blocks.usageBilling")}>
-                <DetailRow label={t("fields.billing")} value={`${formatTooltipUsageCost(detail.item.billedUSD, billingDisplay)} ${detail.item.isFreeModel ? `(${usageLabels.freeModelNoBilling})` : ""}`} />
-                <DetailRow label={t("fields.balanceAfter")} value={formatUsageBalance(detail.item.balanceAfterUSD, billingDisplay)} />
-                <DetailRow label={t("fields.totalTokens")} value={formatCount(usageTotalTokens(detail.item), locale)} mono />
-                <DetailRow label={usageLabels.input} value={formatCount(detail.item.inputTokens, locale)} mono />
-                <DetailRow label={usageLabels.cacheRead} value={formatCount(detail.item.cacheReadTokens, locale)} mono />
-                <DetailRow label={usageLabels.billingDisplay.cacheWrite} value={formatCount(detail.item.cacheWriteTokens, locale)} mono />
-                <DetailRow label={usageLabels.output} value={formatCount(detail.item.outputTokens, locale)} mono />
-                <DetailRow label={t("fields.reasoning")} value={formatCount(detail.item.reasoningTokens, locale)} mono />
-                <DetailRow label={t("fields.callCount")} value={formatCount(detail.item.callCount, locale)} mono />
-                <DetailRow label={t("fields.latency")} value={`${formatCount(detail.item.latencyMS, locale)} ms`} mono />
-              </DetailBlock>
-            </>
-          ) : null}
-
-          {detail?.kind === "usage" ? (
-            <section className="space-y-2">
-              <div className="flex items-center justify-between gap-3 px-1">
-                <h4 className="text-xs font-medium text-foreground/88">{t("rawUsageJsonTitle")}</h4>
-                <CopyActionButton
-                  type="button"
-                  variant="ghost"
-                  size="sm"
-                  className="h-7 px-2 text-xs shadow-none"
-                  value={rawUsageJSON}
-                  messages={copyMessages}
-                  copyOptions={{ copied: t("copied", { label: t("rawUsageJsonTitle") }) }}
-                >
-                  JSON
-                </CopyActionButton>
-              </div>
-              <pre className="max-h-[240px] overflow-auto rounded-lg border border-border/60 bg-muted/35 p-3 text-xs leading-5 text-foreground/86">
-                <code>{rawUsageJSON}</code>
-              </pre>
-            </section>
-          ) : null}
-
-          {detail?.kind === "order" ? (
-            <>
-              <DetailBlock title={t("blocks.order")}>
-                <DetailRow label="ID" value={detail.item.id} mono />
-                <DetailRow label={t("fields.orderNo")} value={detail.item.orderNo} mono />
-                <DetailRow label={t("fields.orderType")} value={detail.item.orderType} />
-                <DetailRow label={t("fields.provider")} value={detail.item.provider} />
-                <DetailRow label={t("fields.status")} value={detail.item.status} />
-                <DetailRow label={t("fields.createdAt")} value={formatDateTime(detail.item.createdAt, locale)} />
-                <DetailRow label={t("fields.paidAt")} value={formatDateTime(detail.item.paidAt, locale)} />
-              </DetailBlock>
-              <DetailBlock title={t("blocks.user")}>
-                <DetailRow label={t("fields.user")} value={resolveUserDisplayName(detail.item.userLabel, detail.item.username, detail.item.userID)} />
-                <DetailRow label={t("fields.userID")} value={detail.item.userID} mono />
-              </DetailBlock>
-              <DetailBlock title={t("blocks.payment")}>
-                <DetailRow label={t("fields.amount")} value={`${formatMoneyCents(detail.item.payAmountCents, detail.item.payCurrency)} / ${formatMoneyCents(detail.item.baseAmountCents, detail.item.baseCurrency)}`} mono />
-                <DetailRow label={t("fields.credit")} value={formatTooltipUsageCost(detail.item.creditUSD, billingDisplay)} mono />
-                <DetailRow label={t("fields.interval")} value={`${detail.item.billingInterval || "-"} x ${detail.item.cycles || 0}`} />
-                <DetailRow label={t("fields.externalPaymentID")} value={detail.item.externalPaymentID || "-"} mono />
-                <DetailRow label={t("fields.externalCheckoutID")} value={detail.item.externalCheckoutID || "-"} mono />
-              </DetailBlock>
-            </>
-          ) : null}
-
-          {detail?.kind === "conversation" ? (
-            <>
-              <DetailBlock title={t("blocks.conversationEvent")}>
-                <DetailRow label="ID" value={detail.item.id} mono />
-                <DetailRow label={t("fields.runID")} value={detail.item.runID} mono />
-                <DetailRow label={t("fields.eventScope")} value={detail.item.eventScope} />
-                <DetailRow label={t("fields.event")} value={detail.item.eventType} />
-                <DetailRow label={t("fields.status")} value={detail.item.status} />
-                <DetailRow label={t("fields.stage")} value={detail.item.stage || detail.item.phase || "-"} />
-                <DetailRow label={t("fields.seq")} value={detail.item.seq} mono />
-                <DetailRow label={t("fields.createdAt")} value={formatDateTime(detail.item.createdAt, locale)} />
-              </DetailBlock>
-              <DetailBlock title={t("blocks.user")}>
-                <DetailRow label={t("fields.user")} value={resolveUserDisplayName(detail.item.userLabel, detail.item.username, detail.item.userID)} />
-                <DetailRow label={t("fields.userID")} value={detail.item.userID} mono />
-                <DetailRow label={t("fields.conversationID")} value={detail.item.conversationID} mono />
-                <DetailRow label={t("fields.messageID")} value={detail.item.messageID} mono />
-              </DetailBlock>
-              <DetailBlock title={t("blocks.modelRoute")}>
-                <DetailRow label={t("fields.platformModel")} value={detail.item.platformModelName || "-"} mono />
-                <DetailRow label={t("fields.upstreamName")} value={detail.item.upstreamName || "-"} />
-                <DetailRow label={t("fields.upstreamModel")} value={detail.item.upstreamModelName || "-"} mono />
-                <DetailRow label={t("fields.bindingCode")} value={detail.item.routedBindingCode || "-"} mono />
-                <DetailRow label={t("fields.protocol")} value={detail.item.providerProtocol || "-"} />
-              </DetailBlock>
-              <DetailBlock title={t("blocks.tool")}>
-                <DetailRow label={t("fields.toolName")} value={detail.item.toolName || "-"} />
-                <DetailRow label={t("fields.toolCallID")} value={detail.item.toolCallID || "-"} mono />
-                <DetailRow label={t("fields.latency")} value={`${formatCount(detail.item.latencyMS, locale)} ms`} mono />
-                <DetailRow label={t("fields.title")} value={detail.item.title || "-"} />
-                <DetailRow label={t("fields.summary")} value={detail.item.summary || "-"} />
-                <DetailRow label={t("fields.payloadSize")} value={formatBytes(detail.item.payloadSizeBytes)} mono />
-              </DetailBlock>
-            </>
-          ) : null}
-
-          <section className="space-y-2">
-            <div className="flex items-center justify-between gap-3 px-1">
-              <h4 className="text-xs font-medium text-foreground/88">{t("jsonTitle")}</h4>
-              <div className="flex items-center gap-1">
-                {requestID ? (
-                  <CopyActionButton
-                    type="button"
-                    variant="ghost"
-                    size="sm"
-                    className="h-7 px-2 text-xs shadow-none"
-                    value={requestID}
-                    messages={copyMessages}
-                    copyOptions={{ copied: t("copied", { label: t("fields.requestID") }) }}
-                  >
-                    {t("fields.requestID")}
-                  </CopyActionButton>
-                ) : null}
-                <CopyActionButton
-                  type="button"
-                  variant="ghost"
-                  size="sm"
-                  className="h-7 px-2 text-xs shadow-none"
-                  value={formattedJSON}
-                  disabled={conversationDetailLoading || (detail?.kind === "conversation" && detail.item.payloadOmitted)}
-                  messages={copyMessages}
-                  copyOptions={{ copied: t("copied", { label: t("jsonTitle") }) }}
-                >
-                  JSON
-                </CopyActionButton>
-              </div>
-            </div>
-            {conversationDetailLoading ? (
-              <div className="flex h-28 items-center justify-center rounded-lg border border-border/60 bg-muted/20">
-                <Spinner label={t("loading")} className="size-4 text-muted-foreground" />
-              </div>
-            ) : (
-              <>
-                {detail?.kind === "conversation" && detail.item.payloadOmitted ? (
-                  <p className="rounded-lg border border-border/60 bg-muted/30 px-3 py-2 text-xs leading-5 text-muted-foreground">
-                    {t("payloadOmitted", { size: formatBytes(detail.item.payloadSizeBytes) })}
-                  </p>
-                ) : null}
-                <pre className="max-h-[320px] overflow-auto rounded-lg border border-border/60 bg-muted/35 p-3 text-xs leading-5 text-foreground/86">
-                  <code>{formattedJSON}</code>
-                </pre>
-              </>
-            )}
-          </section>
-        </div>
-      </SheetContent>
-    </Sheet>
-  );
-}
 
 function AuditLogTable({ onOpenDetail }: { onOpenDetail: (item: AdminAuditLogDTO) => void }) {
   const locale = useLocale();
