+++ frontend/tsconfig.json
@@ -8,10 +8,7 @@
     ],
     "allowJs": false,
     "skipLibCheck": true,
-    "strict": false,
-    "noImplicitAny": true,
-    "strictNullChecks": true,
-    "useUnknownInCatchVariables": true,
+    "strict": true,
     "noEmit": true,
     "incremental": true,
     "module": "esnext",
